Create rdbuf.cpp
The getline() and get() functions require a variable to store its content. If all you want to do is to read a file however, then the variable becomes useful and simply takes up extra memory. When a file is opened, its content is buffered or stored in memory until it gets closed. During this buffering stage, you can use the function rdbuf() to read the content of the file without having to create any variables.