Create Delimiter.cpp
Delimiters are a predefined character that separates one piece of information from another. Some common delimiters involve white spaces (' ') and commas (','). When using getline() previously, we only made use of two of its parameters. It actually has a third parameter which is reserved for a delimiter. By default, this delimiter is the newline character ('\n'). Thus the parameters for getline(x, y, z) are:
The stream source (x).
The string variable to store what is read (y).
The delimiter to separate the content of the stream source (z).

The code above continues to read the stream and separates the content via the delimiter, ,. This is why there are no commas present in the output.