Create Challenge Task (files).cpp
Write a program that reads a text file . This file is stored in the variable path.
The file contains several instances of the word Burma. Replace each instance of Burma with Myanmar, and print the results of this transformation. The final output of your program should be:

Myanmar is a country in Southeast Asia.
The capital of Myanmar is Naypyidaw.
Its population is about 54 million people.
Myanmar is a former British colony.